refactor: rename browser_run.py to loaders.py, add notebook with real outputs Address maintainer review feedback: - Rename browser_run.py to loaders.py (matches module naming convention) - Add docs/browser_run.ipynb with executed outputs from real API calls:   - Workers AI docs loaded as markdown (15K chars from JS-rendered page)   - books.toscrape.com crawled for knowledge base ingestion   - cloudflare.com scraped for h1, h2, nav elements   - Pricing data extracted as structured JSON from cloudflare.com/plans   - Schema-enforced company extraction from what-is-cloudflare page   - 88 links discovered from Browser Run docs   - Screenshot captured (398K base64 PNG)   - Full research pipeline: discover links -> load 3 pages -> summarize with Llama 3.3 70B - LangGraph patterns shown as notebook examples per maintainer guidance
